+++ resources/views/modules/landing/layouts/app.blade.php
@@ -150,8 +150,8 @@
                     : '/assets/landing/images/berita/default.jpg';
 
                 const date = news.published_at
-                    ? new Date(news.published_at).toLocaleDateString('id-ID', { day: '2-digit', month: 'short', year: 'numeric' })
-                    : new Date(news.created_at).toLocaleDateString('id-ID', { day: '2-digit', month: 'short', year: 'numeric' });
+                    ? new Date(news.published_at).toLocaleDateString('id-ID', { day: '2-digit', month: 'long', year: 'numeric' })
+                    : new Date(news.created_at).toLocaleDateString('id-ID', { day: '2-digit', month: 'long', year: 'numeric' });
 
                 const newsUrl = news.slug ? `/berita/${news.slug}` : `/berita/${news.id}`;
